Adds Modifiers & Advantages from GURPS Psionic Powers (#107)
Includes:
Limitations and enhancements to Neutralize and static, applied to the advantages in both powers and Basic set. Also Hallucinations, to Powers
Also includes new limitation and enhancement modifiers, as well as adding the new advantages from Psionic Powers to the Psionic Powers advantages library.

For some reason saving this has made GCS re-order the references creating a lot of needless changes but I guess that's just the way it is.